refactor(subscribe): use form element `action` and `method` to `POST` form when JS is disabled
