+++ src/lib/components/Subscribe/_styles.scss
@@ -1,0 +1,4 @@
+web-subscribe {
+  display: block;
+  width: 100%;
+}
