--- src/lib/components/SubscribeForm/_styles.scss
@@ -1,75 +1,0 @@
-@import '../../../styles/settings/colors';
-@import '../../../styles/tools/breakpoints';
-
-web-subscribe {
-  display: block;
-  margin-top: 32px;
-  text-align: center;
-  width: 100%;
-
-  .web-subscribe__border {
-    border: 2px solid #F1F3F4;
-    border-top-left-radius: 2px;
-    border-top-right-radius: 2px;
-    width: 100%;
-  }
-
-  .web-subscribe__button {
-    margin-left: auto;
-    margin-right: auto;
-  }
-
-  .web-subscribe__error {
-    color: $WARNING_COLOR;
-  }
-
-  .web-subscribe__field {
-    display: flex;
-    justify-content: space-between;
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: 8px;
-    text-align: left;
-  }
-
-  .web-subscribe__form {
-    margin-left: auto;
-    margin-right: auto;
-    max-width: 720px;
-    width: calc(100% - 32px);
-  }
-
-  .web-subscribe__input {
-    border: 2px solid #F1F3F4;
-    border-top-left-radius: 2px;
-    border-top-right-radius: 2px;
-    font: inherit;
-    font-size: 16px;
-    line-height: 20px;
-    padding: 8px 8px 8px 8px;
-    width: 100%;
-
-    &:focus {
-      box-shadow: 0 2px 0 0 $WEB_SECONDARY_COLOR;
-      outline: none;
-    }
-  }
-
-  .web-subscribe__select {
-    background: white;
-    border: 0;
-    font: inherit;
-    height: 36px;
-  }
-
-  @include bp(sm) {
-    .web-subscribe__fields {
-      display: flex;
-      justify-content: space-between;
-
-      .web-subscribe__field {
-        width: calc(50% - 4px);
-      }
-    }
-  }
-}
