+++ src/site/_data/site.js
@@ -25,6 +25,8 @@ module.exports = {
   url: "https://web.dev",
   repo: "https://github.com/GoogleChrome/web.dev",
   subscribe: "https://web.dev/subscribe",
+  subscribeForm:
+    "https://services.google.com/fb/submissions/591768a1-61a6-4f16-8e3c-adf1661539da/",
   thumbnail: "/images/social.png",
   isBannerEnabled: false,
   banner: ``,
