+++ src/site/_includes/partials/post.njk
@@ -136,5 +136,5 @@
       {% endif %} {% endfor %}
     </ul>
   </div> #}
-  <web-subscribe></web-subscribe>
+  {% include 'partials/subscribe.njk' %}
 </div>
