+++ src/styles/components/_all.scss
@@ -29,6 +29,7 @@
 @import 'progress';
 @import 'self-assessment-hint';
 @import 'stats';
+@import 'subscribe';
 @import 'topic';
 @import 'type';
 @import 'tooltips';
